pyproject.toml
[project]
name = "conference-site"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi[standard]>=0.115.12",
    "matplotlib>=3.10.3",
    "openpyxl>=3.1.5",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.8.1",
    "python-jose>=3.4.0",
    "sqlalchemy>=2.0.40",
    "sqlmodel>=0.0.24",
]

[dependency-groups]
dev = [
    "alembic>=1.15.2",
    "black>=25.1.0",
    "djlint>=1.36.4",
    "greenlet>=3.1.1",
    "mypy>=1.15.0",
    "python-lsp-server[all]>=1.12.2",
    "ruff>=0.11.2",
]

[tool.djlint]
ignore="H006,H013"

[[tool.mypy.overrides]]
module = ["jose.*"]
follow_untyped_imports = true

[[tool.mypy.overrides]]
module = ["passlib.*"]
follow_untyped_imports = true

[tool.ruff]
line-length = 79

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]